Push screenshot to separate file.
