+++ client/my-sites/theme/theme-sheet-content/index.jsx
@@ -9,45 +9,11 @@ import React from 'react';
  * Internal dependencies
  */
 import Gridicon from 'components/gridicon';
-import { localize } from 'i18n-calypso';
 import Navigation from './theme-content-sections/navigation';
 import Overview from './theme-content-sections/overview';
 import Setup from './theme-content-sections/setup';
 import Support from './theme-content-sections/support';
-
-const PreviewButton = localize(
-	( { theme, togglePreview, translate } ) => {
-		if ( ! theme.demo_uri ) {
-			return null;
-		}
-
-		return (
-			<a className="theme__sheet-preview-link" onClick={ togglePreview } data-tip-target="theme-sheet-preview">
-				<Gridicon icon="themes" size={ 18 } />
-				<span className="theme__sheet-preview-link-text">
-					{ translate( 'Open Live Demo', { context: 'Individual theme live preview button' } ) }
-				</span>
-			</a>
-		);
-	}
-);
-
-const Screenshot = ( { isLoaded, isJetpack, theme, togglePreview } ) => {
-	const fullLengthScreenshot = () =>
-		isLoaded ? theme.screenshots[ 0 ] : null;
-	const screenshot = isJetpack ? theme.screenshot : fullLengthScreenshot();
-	const img = screenshot && <img className="theme__sheet-img" src={ screenshot + '?=w680' } />;
-
-	return (
-		<div className="theme__sheet-screenshot">
-			<PreviewButton
-				togglePreview={ togglePreview }
-				theme={ theme }
-			/>
-			{ img }
-		</div>
-	);
-};
+import Screenshot from './theme-content-sections/screenshot'
 
 class ThemeSheetContent extends React.Component {
 	static propTypes = {
@@ -70,7 +36,7 @@ class ThemeSheetContent extends React.Component {
 	}
 
 	render() {
-		const { id, isJetpack, isLoaded, theme, section, siteSlug } = this.props;
+		const { id, isJetpack, isLoaded, theme, section, siteSlug, togglePreview } = this.props;
 
 		return (
 			<div className="theme__sheet-columns">
@@ -92,6 +58,7 @@ class ThemeSheetContent extends React.Component {
 						isJetpack={ isJetpack }
 						isLoaded={ isLoaded }
 						theme={ theme }
+						togglePreview={ togglePreview }
 					/>
 				</div>
 			</div>
@@ -99,4 +66,4 @@ class ThemeSheetContent extends React.Component {
 	}
 }
 
-export default localize( ThemeSheetContent );
+export default ThemeSheetContent;
